Moved ITransaction.h to lib/CryptoNoteCore
