+++ include/ITransfersContainer.h
@@ -22,9 +22,9 @@
 #include <limits>
 #include <vector>
 #include <crypto/hash.h>
+#include <CryptoNoteCore/ITransaction.h>
 #include <IObservable.h>
 #include <IStreamSerializable.h>
-#include <ITransaction.h>
 
 namespace CryptoNote {
