+++ lib/CryptoNoteCore/ITransaction.h
