+++ lib/Transfers/TransfersContainer.h
@@ -30,10 +30,10 @@
 #include <CryptoNoteCore/CryptoNoteBasic.h>
 #include <CryptoNoteCore/CryptoNoteSerialization.h>
 #include <CryptoNoteCore/Currency.h>
+#include <CryptoNoteCore/ITransaction.h>
 #include <Logging/LoggerRef.h>
 #include <Serialization/ISerializer.h>
 #include <Serialization/SerializationOverloads.h>
-#include <ITransaction.h>
 #include <ITransfersContainer.h>
 
 namespace CryptoNote {
